app_server/requirements.txt
pandas
numpy
Flask
Flask-Cors
psycopg2-binary
SQLAlchemy
flask-sqlalchemy
requests
